+++ src/SDL_renderer_textengine.c
@@ -853,7 +853,7 @@ TTF_TextEngine *TTF_CreateRendererTextEngineWithProperties(SDL_PropertiesID prop
         return NULL;
     }
 
-    int atlas_texture_size = SDL_GetNumberProperty(props, TTF_PROP_RENDERER_TEXT_ENGINE_ATLAS_TEXTURE_SIZE, 512);
+    int atlas_texture_size = (int)SDL_GetNumberProperty(props, TTF_PROP_RENDERER_TEXT_ENGINE_ATLAS_TEXTURE_SIZE, 512);
     if (atlas_texture_size <= 0) {
         SDL_SetError("Failed to create renderer text engine: Invalid texture atlas size.");
         return NULL;
